opentinker/requirements.txt
# OpenTinker Requirements
# Python 3.8+

# Core dependencies
ray>=2.9.0
torch>=2.0.0
transformers>=4.35.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Configuration
omegaconf>=2.3.0
hydra-core>=1.3.0
pyyaml>=6.0

# Data processing
pandas>=2.0.0
pyarrow>=14.0.0
datasets>=2.14.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0

# Optional: Logging and monitoring
wandb>=0.16.0

# Optional: Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
